# PERF: copy cached attributes on extension index shallow_copy