--- a/pandas/core/indexes/period.py
+++ b/pandas/core/indexes/period.py
@@ -233,6 +233,7 @@ class PeriodIndex(DatetimeIndexOpsMixin, Int64Index):
         # For groupby perf. See note in indexes/base about _index_data
         result._index_data = values._data
         result.name = name
+        result._cache = {}
         result._reset_identity()
         return result